+++ js/utils/utilities.js
@@ -710,6 +710,12 @@ var Utils = {
 
     parseNumber: function(val, thousand, decimal){
         return val.replace(new RegExp('\\'+thousand, "g"), "").replace(new RegExp('\\'+decimal, 'g'), ".");
+    },
+
+    nearest: function(val, precision, down){
+        val /= precision;
+        val = Math[down === true ? 'floor' : 'ceil'](val) * precision;
+        return val;
     }
 };